FEAT: Added Kaggle API raw data stream and designated it for research
ABOUT (Brief)
Following files were created:
- ./src/collect_raw_data.py
- ./requirements.txt
- ./data/raw_data.csv

ABOUT (Detailed)
The research requires a single raw dataset available on Kaggle. Thus, a collection process for raw data and streaming was developed as a simple implementation of the Kaggle API. This process features automated steps for unzipping the downloaded ZIP file, renaming the dataset to `raw_data.csv`, and placing it in the appropriate directory, `./data/`. The following files were created:
- Added ./src/collect_raw_data.py to collect raw data
- Added ./requirements.txt with actual dependencies
- Added ./data/raw_data.csv with collected raw data